Add venues api and adapters.

+++ core/src/main/java/com/place/search/core/models/PlaceCoordinates.kt
@@ -1,3 +1,3 @@
 package com.place.search.core.models
 
-data class PlaceCoordinates(val lat: Long, val lng: Long)
+data class PlaceCoordinates(val lat: Double, val lng: Double)
